end v2

+++ Разработка/Backend/first_player/first_player/Controllers/AdminController.cs
@@ -25,7 +25,7 @@ namespace first_player.Controllers
         [HttpDelete("{nickname}")]
         public async Task<IActionResult> Delete(string nickname)
         {
-            var userToDelete = await db_people.Users.FirstOrDefaultAsync(u => u.nickname == nickname);
+            var userToDelete = db_people.Users.FirstOrDefault(u => u.nickname == nickname);
 
             if (userToDelete == null)
             {
@@ -33,17 +33,31 @@ namespace first_player.Controllers
             }
 
             // Удаление связанных комментариев
-            var userComments = await db_people.Comments.Where(c => c.UsersId == userToDelete.Id).ToListAsync();
-            db_people.Comments.RemoveRange(userComments);
-
+            var userComments =  db_people.Comments.Where(c => c.UsersId == userToDelete.Id).ToList();
+            if(userComments != null)
+            {
+                db_people.Comments.RemoveRange(userComments);
+                db_people.SaveChanges();
+            }
             // Удаление связанных постов
-            var userPosts = await db_people.Posts.Where(p => p.UsersId == userToDelete.Id).ToListAsync();
-            db_people.Posts.RemoveRange(userPosts);
-
+            var userPosts =  db_people.Posts.Where(p => p.UsersId == userToDelete.Id).ToList();
+            foreach (var post in userPosts)
+            {
+                // Найти все комментарии, принадлежащие этому посту
+                var postComments = db_people.Comments.Where(c => c.PostsId == post.Id).ToList();
+                if (postComments != null)
+                {
+                    db_people.Comments.RemoveRange(postComments);
+                }
+            }
+            if (userPosts != null)
+            {
+                db_people.Posts.RemoveRange(userPosts);
+                db_people.SaveChanges();
+            }
             // Удаление пользователя
-            db_people.Users.Remove(userToDelete);
-
-            await db_people.SaveChangesAsync();
+            db_people.Users.RemoveRange(userToDelete);
+            db_people.SaveChanges();
 
             return Ok("Пользователь успешно удален");
         }
